backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
gunicorn==23.0.0
Pillow==11.0.0
psycopg[binary]==3.2.3
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
whitenoise==6.7.0
